refactor: metadata repository
